Use bevy_simple_prefs for persisting settings
