+++ Cargo.toml
@@ -18,6 +18,7 @@ strum = "0.27.1"
 strum_macros = "0.27.1"
 grid = "0.17.0"
 thiserror = "1.0"
+bevy_simple_prefs = "0.6.0"
 
 [features]
 inspector = ["bevy-inspector-egui"]
